Started the spectrometer code in earnest.
This code carries over the basic UART and command-interpreter code
and adds the SPI data-exchange functions.

The format for the interpreter command is:
X <csi> <byte0> <byte1> ... Ctrl-J
where <csi> is an integer representing the chip to select and
the bytes are in base 10 representation.

The response is shown in hexadecimal.